WIP refactor part 2
