+++ flask_monitoringdashboard/templates/dashboard/dashboard-base.html
@@ -1,39 +1,46 @@
 {% extends "base.html" %}
-{% set title="Measurement Overview" %}
+{% set title="Dashboard Overview" %}
 {% block content %}
 
 {% block menu %}
+{% set url_overview='dashboard.overview' %}
+{% set url_heatmap='dashboard.heatmap' %}
+{% set url_version_usage='dashboard.version_usage' %}
+{% set url_req_endpoint='dashboard.page_number_of_requests_per_endpoint' %}
+{% set url_boxplot_version='dashboard.page_boxplot_per_version' %}
+{% set url_boxplot_endpoint='dashboard.page_boxplot_per_endpoint' %}
+
 <nav class="navbar navbar-expand-sm navbar-dark" style="background: #343a40; margin-bottom: 15px;">
   <div class="collapse navbar-collapse" id="nav-content">
-    <ul class="nav">
+    <ul class="navbar-nav nav-pills">
       <li class="nav-item">
-        <a class="nav-link" style="color: white" href="{{ url_for('dashboard.overview') }}">
-          Overview
+        <a class="nav-link {{ active_if_is(url_overview) }}" href="{{ url_for(url_overview) }}">
+            <span class="nav-link-text">Overview</span>
         </a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" style="color: white" href="{{ url_for('dashboard.heatmap') }}">
-           Heatmap of number of requests
+        <a class="nav-link {{ active_if_is(url_heatmap) }}" href="{{ url_for(url_heatmap) }}">
+          <span class="nav-link-text">Heatmap of number of requests</span>
         </a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" style="color: white" href="{{ url_for('dashboard.version_usage') }}">
-          Version usage
+        <a class="nav-link {{ active_if_is(url_version_usage) }}" href="{{ url_for(url_version_usage) }}">
+            <span class="nav-link-text">Version usage</span>
         </a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" style="color: white" href="{{ url_for('dashboard.page_number_of_requests_per_endpoint') }}">
-          Requests per endpoint
+        <a class="nav-link {{ active_if_is(url_req_endpoint) }}" href="{{ url_for(url_req_endpoint) }}">
+            <span class="nav-link-text">Requests per endpoint</span>
         </a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" style="color: white" href="{{ url_for('dashboard.page_boxplot_per_version') }}">
-          Time per version
+        <a class="nav-link {{ active_if_is(url_boxplot_version) }}" href="{{ url_for(url_boxplot_version) }}">
+            <span class="nav-link-text">Time per version</span>
         </a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" style="color: white" href="{{ url_for('dashboard.page_boxplot_per_endpoint') }}">
-          Time per endpoint
+        <a class="nav-link {{ active_if_is(url_boxplot_endpoint) }}" href="{{ url_for(url_boxplot_endpoint) }}">
+            <span class="nav-link-text">Time per endpoint</span>
         </a>
       </li>
     </ul>
