+++ flask_monitoringdashboard/templates/dashboard/dashboard.html
@@ -1,4 +1,4 @@
-{% extends "dashboard/measurement-overview.html" %}
+{% extends "dashboard/dashboard-base.html" %}
 
 {% block subcontent %}
 <div class="card mb-3">
