--- flask_monitoringdashboard/templates/dashboard/settings.html
@@ -1,42 +1,0 @@
-{% extends "dashboard/measurement-overview.html" %}
-{% block content %}
-
-<div class="card mb-3">
-  <div class="card-header">Settings</div>
-  <div class="card-body">
-    <div class="table-responsive">
-      <table class="table table-striped table-bordered table-hover">
-                        <tr>
-                            <td><b>Dashboard link</b></td>
-                            <td>/{{ link }}</td>
-                        </tr>
-                        <tr>
-                            <td><b>Version of monitored app</b></td>
-                            <td>{{ config.version }}</td>
-                        </tr>
-                        <tr>
-                            <td><b>Group by variable</b></td>
-                            <td>{{ group_by if group_by != None else "Unspecified" }}</td>
-                        </tr>
-                        <tr>
-                            <td><b>Measurements database</b></td>
-                            <td>{{ config.database_name }}</td>
-                        </tr>
-                        <tr>
-                            <td><b>Location of tests of monitored app</b></td>
-                            <td>{{ config.test_dir if config.test_dir != None else "Unspecified" }}</td>
-                        </tr>
-                        <tr>
-                            <td><b>Username</b></td>
-                            <td>{{ config.username }}</td>
-                        </tr>
-                        <tr>
-                            <td><b>Password</b></td>
-                            <td>{{ pw }}</td>
-                        </tr>
-                    </table>
-                </div>
-            </div>
-        </div>
-
-{% endblock %}
