+++ flask_monitoringdashboard/templates/endpoint/plotly.html
@@ -1,4 +1,4 @@
-{% extends "endpoint/main.html" %}
+{% extends "endpoint/details-base.html" %}
 {% block graph_content %}
     <div class="col-md-12 col-sm-12">
         {{ graph|safe }}
