+++ flask_monitoringdashboard/templates/endpoint/time_per_user.html
@@ -1,4 +1,4 @@
-{% extends "endpoint/main.html" %}
+{% extends "endpoint/details-base.html" %}
 {% block graph_content %}
     <p>Select a number of items below to reduce the size of the graph.</p>
     <br/>
