+++ flask_monitoringdashboard/templates/export-data.html
@@ -1,4 +1,4 @@
 {% for entry in data %}
-    {{ entry }}
-    <br>
+  {{ entry }}
+  <br>
 {% endfor %}
