+++ flask_monitoringdashboard/templates/login.html
@@ -3,16 +3,16 @@
 <body>
 <div class="container py-5">
   <div class="row">
-    <div class="col-md-6 offset-md-3">
-      <div class="text-center">
+    <div class="col-md-8 offset-md-2">
+      <div class="text-center" style="margin-bottom: 50px;">
         <h1>Flask Monitoring Dashboard</h1>
-        <hr />
+        <span>Automatically monitor the evolving performance of Flask/Python web services</span>
       </div>
     </div>
   </div>
 
   <div class="row">
-    <div class="col-md-6 offset-md-3">
+    <div class="col-md-4 offset-md-4">
       <span class="anchor" id="formLogin"></span>
       <div class="card card-outline-secondary">
         <div class="card-header">
@@ -30,15 +30,15 @@
               {{ form.hidden_tag() }}
             <div class="form-group">
                 {{ form.name.label }}
-                {{ form.name(class_="form-control") }}
+                {{ form.name(class_="form-control", placeholder="Username", autofocus=true, required=true) }}
             </div>
 
             <div class="form-group">
               {{ form.password.label }}
-              {{ form.password(class_="form-control") }}
+              {{ form.password(class_="form-control", placeholder="Password", required=true) }}
             </div>
 
-            {{ form.submit(class_="btn btn-primary btn-lg") }}
+            {{ form.submit(class_="btn btn-primary btn-block") }}
           </form>
           <hr />
