+++ flask_monitoringdashboard/templates/rules.html
@@ -1,8 +1,6 @@
-{% extends "dashboard/measurement.html" %}
-{% block subcontent %}
-{% block menu %}{% endblock %}
+{% extends "base.html" %}
 {% set title="All Rules" %}
-
+{% block content %}
 <div class="card mb-3">
   <div class="card-header">{{ title }}</div>
   <div class="card-body">
